fix(muc): Encrypted messages are not sending correctly (#37)
ver: 0.5.11

fix #36

feat(exec): Add purge device list command

refactor(db)!: Reset schema numeration due to invalid migrations

+++ core/mock/net/src/main/java/cc/cryptopunks/crypton/mock/MockChatNet.kt
@@ -16,7 +16,7 @@ class MockChatNet(
         TODO("Not yet implemented")
     }
 
-    override fun joinConference(address: Address, nickname: String, historySince: Int) {
+    override fun joinConference(address: Address, nickname: String, historySince: Long) {
         TODO("Not yet implemented")
     }
